Allow render_email to be used outside nereid context
render_email method required a nereid request context to work since it
used the render_template functionality of nereid.templating. This patch
allows jinja2.Template instances to be sent as text and html templates
in which case, these templates are rendered with the context, instead of
using nereid.template.render_template. This is useful if you want to
call render_email for example from a method in a request to trytond.