Remove SiteNamePrefix Loader #2853
The template loader design with SiteNamePrefix loader works once the
name of the template is passed and not for dynamically changing the
name. This is because jinja caching (environment _load_template) uses
the name to cache the content and the second time the template is loaded
the name would not have the site prefix.