Add CSRF protection through Flask-WTF
CSRF protection was optional until now, and the developer had
to use his favorite tool to get the protection. Since most apps
in practice never implemented this, it was suggested that the
protection also be included with Nereid.

The implementation uses Flask-WTF, a very popular library in
the flask ecosystem for handling forms and CSRF without re-inventing
the wheel