Avoid stale AR being reused
The patch that fixed the retry api introduced a case
where the active record tied to the request could have
come from a transaction that is already closed, leaving
an AR that is stale (closed cursor).

This fixes the issue by refactoring the locale lookup to
be a method in website and then calling it in the transaction
where the website is looked up.